% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dataset_path}
\alias{dataset_path}
\title{dataset_path}
\usage{
dataset_path(overture_type)
}
\arguments{
\item{overture_type}{Character. Required. The type of feature to select.
Examples include 'building', 'place', etc. To learn more, run \code{get_all_overture_types()}.}
}
\value{
Character. The S3 path to the bucket where the data is stored.
}
\description{
This function returns the S3 path for the specified Overture dataset type.
}
\examples{
# Example usage
path <- dataset_path('place')
print(path)
}
