% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_all_overture_schema_types}
\alias{get_all_overture_schema_types}
\title{get_all_overture_types}
\usage{
get_all_overture_schema_types()
}
\value{
Character vector. All overturemaps theme types.
}
\description{
This function returns all of the overturemaps theme types.
}
\note{
The theme types are important for fetching data from the S3 bucket,
as they indicate if you are fetching places, buildings, admin, etc.
}
\examples{
# Example usage
types <- get_all_overture_schema_types()
print(types)

}
